--- src/main/java/duke/command/CommandResponse.java
@@ -1,13 +1,0 @@
-package duke.command;
-
-public class CommandResponse {
-    private String[] response;
-
-    CommandResponse(String... response) {
-        this.response = response;
-    }
-
-    public String[] get() {
-        return response;
-    }
-}
