--- src/main/java/duke/command/ExitCommand.java
@@ -1,17 +1,0 @@
-package duke.command;
-
-import duke.task.TaskList;
-
-import static duke.common.Messages.GOODBYE_MSG;
-
-public class ExitCommand implements Command {
-    @Override
-    public boolean isExitCommand() {
-        return true;
-    }
-
-    @Override
-    public CommandResponse execute(TaskList taskList) {
-        return new CommandResponse(GOODBYE_MSG);
-    }
-}
