--- src/main/java/duke/common/DateTimeFormat.java
@@ -1,8 +1,0 @@
-package duke.common;
-
-import java.time.format.DateTimeFormatter;
-
-public class DateTimeFormat {
-    public static final DateTimeFormatter INPUT_DATE_TIME_FORMAT = DateTimeFormatter.ofPattern("d/M/yyyy H:m");
-    public static final DateTimeFormatter OUTPUT_DATE_TIME_FORMAT = DateTimeFormatter.ofPattern("dd MMM yyyy, hh:mma");
-}
