--- src/main/java/duke/exception/DukeException.java
@@ -1,14 +1,0 @@
-package duke.exception;
-
-public abstract class DukeException extends Exception {
-    protected final String[] messages;
-
-    public DukeException(String... messages) {
-        super("An error occurred in the program.");
-        this.messages = messages;
-    }
-
-    public String[] getMessages() {
-        return messages;
-    }
-}
