+++ src/main/java/seedu/duke/command/CommandResponse.java
@@ -1,0 +1,13 @@
+package seedu.duke.command;
+
+public class CommandResponse {
+    private String[] response;
+
+    CommandResponse(String... response) {
+        this.response = response;
+    }
+
+    public String[] get() {
+        return response;
+    }
+}
