+++ src/main/java/seedu/duke/command/ExitCommand.java
@@ -1,0 +1,16 @@
+package seedu.duke.command;
+
+import seedu.duke.common.Messages;
+import seedu.duke.task.TaskList;
+
+public class ExitCommand implements Command {
+    @Override
+    public boolean isExitCommand() {
+        return true;
+    }
+
+    @Override
+    public CommandResponse execute(TaskList taskList) {
+        return new CommandResponse(Messages.GOODBYE_MSG);
+    }
+}
