+++ src/main/java/seedu/duke/common/DateTimeFormat.java
@@ -1,0 +1,8 @@
+package seedu.duke.common;
+
+import java.time.format.DateTimeFormatter;
+
+public class DateTimeFormat {
+    public static final DateTimeFormatter INPUT_DATE_TIME_FORMAT = DateTimeFormatter.ofPattern("d/M/yyyy H:m");
+    public static final DateTimeFormatter OUTPUT_DATE_TIME_FORMAT = DateTimeFormatter.ofPattern("dd MMM yyyy, hh:mma");
+}
