+++ src/main/java/seedu/duke/exception/InvalidStorageFilePathException.java
@@ -1,0 +1,12 @@
+package seedu.duke.exception;
+
+public class InvalidStorageFilePathException extends DukeException {
+    /**
+     * Constructs a new InvalidStorageFilePathException with the specified array of messages.
+     *
+     * @param messages The detail array of messages.
+     */
+    public InvalidStorageFilePathException(String... messages) {
+        super(messages);
+    }
+}
