+++ src/main/java/seedu/duke/exception/StorageOperationException.java
@@ -1,0 +1,12 @@
+package seedu.duke.exception;
+
+public class StorageOperationException extends DukeException {
+    /**
+     * Constructs a new StorageOperationException with the specified array of messages.
+     *
+     * @param messages The detail array of messages.
+     */
+    public StorageOperationException(String... messages) {
+        super(messages);
+    }
+}
